+++ packages/lib/src/scripts/infinite-grid/infinite-grid.js
@@ -13,6 +13,7 @@ import {
     BLENDMODE_SRC_ALPHA,
     BLENDEQUATION_ADD
 } from 'playcanvas';
+import { warnOnce } from '../../utils/validation.ts';
 
 const vsCode = /* glsl*/ `
     uniform mat4 camera_matrix;
@@ -178,9 +179,23 @@ const attributes = {
     vertex_position: SEMANTIC_POSITION
 };
 
+/**
+ * @deprecated This script is deprecated and will be removed in the next major version.
+ *
+ * Use {@link https://github.com/playcanvas/engine/tree/main/scripts/esm/grid.mjs | Grid} from `playcanvas` instead.
+ *
+ * ```tsx
+ * import { Grid } from "playcanvas/scripts/esm/grid.mjs";
+ *
+ * export const MyScript = () => <Script script={Grid} />;
+ * ```
+ */
 export class Grid extends Script {
+    static scriptName = 'grid';
 
     initialize() {
+
+        warnOnce('This script is deprecated and will be removed in the next major version. Use the `Grid` from `playcanvas` instead.');
 
         const layerName = 'World';