+++ packages/lib/src/scripts/shadow-catcher/index.js
@@ -1,11 +1,24 @@
 import { Script, Entity, Layer, StandardMaterial, BLEND_NORMAL, SHADOW_VSM16, SHADOWUPDATE_REALTIME, CHUNKAPI_2_1 } from 'playcanvas';
+import { warnOnce } from '../../utils/validation.ts';
 
 const endPS = `
     litArgs_opacity = mix(light0_shadowIntensity, 0.0, shadow0);
     gl_FragColor.rgb = vec3(0.0);
 `;
 
+/**
+ * @deprecated This script is deprecated and will be removed in the next major version.
+ * 
+ * Use {@link https://github.com/playcanvas/engine/tree/main/scripts/esm/shadow-catcher.mjs | ShadowCatcher} from `playcanvas` instead.
+ * 
+ * ```tsx
+ * import { ShadowCatcher } from "playcanvas/scripts/esm/shadow-catcher.mjs";
+ *
+ * export const MyScript = () => <Script script={ShadowCatcher} />;
+ * ```
+ */
 export class ShadowCatcher extends Script {
+    static scriptName = 'shadowCatcher';
     /**
      * The shadow distance of the shadow catcher light.
      * @type {number}
@@ -43,6 +56,8 @@ export class ShadowCatcher extends Script {
     light = null;
 
     initialize() {
+        warnOnce('This script is deprecated and will be removed in the next major version. Use the `ShadowCatcher` from `playcanvas` instead.');
+
         // create and add the shadow layer
         this.layer = new Layer({
             name: 'Shadow Layer'